fix: migrate to Odin 2026-03 core library APIs, fix cross-platform builds
The Odin 2026-03 release changed several core library signatures across
sys/linux, sys/posix, sys/windows, and sys/linux/uring. This broke
builds
on Linux, Windows, and (for some files) all non-Darwin targets.
Darwin/BSD
happened to pass because their platform files used fewer affected APIs.

Linux (sys_time, sys_thread, io_backend):
- clock_gettime now returns (Time_Spec, Errno); use .BOOTTIME enum value
- posix.madvise renamed; use foreign libc madvise directly since
  MADV_HUGEPAGE/MADV_POPULATE_WRITE are Linux-specific, not in POSIX
- linux.socket gained a separate Socket_FD_Flags parameter
- linux.bind/connect are now polymorphic (no addr_len arg)
- linux.setsockopt split into typed overloads; use setsockopt_base
- linux.Shut_How renamed to Shutdown_How
- uring.read/write offset changed from i64 to u64
- uring.accept addr param is polymorphic; use typed nil for fallback
- Msg_Hdr.iov is now []IO_Vec; cast via ([^]IO_Vec) for single-element
- pthread_t <-> rawptr casts routed through uintptr (matches BSD impl)
- Added missing import "core:c" for c.int usage

Windows (io_backend, sys_thread, shard, watchdog):
- LPFN_ACCEPTEX/CONNECTEX return BOOL not bool; use win.TRUE/FALSE
- OVERLAPPED vs WSAOVERLAPPED are now distinct structs; cast at WSA
  call sites since the memory layout is ABI-compatible
- WSASendTo/WSARecvFrom/GetAcceptExSockaddrs use ^sockaddr (lowercase)
- win.bind takes ^SOCKADDR_STORAGE_LH directly
- Added import "core:testing" for test procs
- Renamed _setjmp/_longjmp foreign bindings to avoid collision with
  core:c/libc/setjmp.odin declarations
- os_signal_thread changed from c.int to posix.Signal to match POSIX
  platform signatures
- Created sys_signals_windows.odin stub (Sleep-based poll, returns
  no-signal so watchdog heartbeat monitoring still works)
- Guarded SIGUSR2/SIGHUP switch cases with when ODIN_OS != .Windows
  since these map to the same enum value (-1) on Windows

Windows backend DOD improvement (io_backend_windows):
- Introduced Win_Op_Data as struct #raw_union overlapping Accept and
  Recvfrom kernel-writable buffers. These are mutually exclusive per
  entry (discriminated by entry.operation), saving ~96 bytes per entry
  across 512 entries (~48KB total). Improves L2/L3 cache utilization
  on the _backend_collect hot path.
- Added compile-time layout guards (#assert) to catch union mis-sizing
  or AcceptEx buffer capacity drift
- Added TINA_DEBUG_ASSERTS config flag (-define:TINA_ASSERTS=true) for
  opt-in runtime union variant verification without production cost

Test stack overflow fixes:
- Converted Reactor/Shard/Platform_Backend from stack locals to
  heap-allocated pointers in test procs. Production code already uses
  new(Shard) and Grand Arena; only tests were stack-allocating these
  ~300KB+ structs, triggering Windows stack overflow warnings.

All 12 supported targets now builds cleanly:
darwin_{arm64,amd64}, linux_{amd64,arm64,i386,arm32},
windows_amd64, freebsd_{amd64,arm64}, openbsd_amd64,
netbsd_{amd64,arm64}

+++ src/sys_thread_posix.odin
@@ -77,15 +77,16 @@ when ODIN_OS == .Darwin {
 		pthread_set_name_np(pthread_self(), cast(cstring)&buf[0])
 	}
 
-	// Returns the OS-specific thread handle (pthread_t on POSIX)
+	// Returns the OS-specific thread handle (pthread_t on POSIX).
+	// On FreeBSD, pthread_t is u64 (not rawptr), so we route through uintptr.
 	os_get_current_thread_handle :: proc "contextless" () -> rawptr {
-		return rawptr(pthread_self())
+		return rawptr(uintptr(pthread_self()))
 	}
 
-	// Sends a signal directly to a specific thread (for Watchdog forced recovery)
+	// Sends a signal directly to a specific thread (for Watchdog forced recovery).
+	// Reverses the uintptr packing from os_get_current_thread_handle.
 	os_signal_thread :: proc "contextless" (thread_handle: rawptr, sig: posix.Signal) {
-		// pthread_kill requires the pthread_t and the signal number
-		posix.pthread_kill(posix.pthread_t(thread_handle), sig)
+		posix.pthread_kill(posix.pthread_t(uintptr(thread_handle)), sig)
 	}
 }
